capacity and weight must be int not floats

+++ daliuge-translator/dlg/dropmake/utils/antichains.py
@@ -44,9 +44,9 @@ def _create_split_graph(dag, w_attr='weight'):
         xi = '{0}_x'.format(el[0])
         yi = '{0}_y'.format(el[0])
         #print(el)
-        bpg.add_edge('s', xi, capacity=float(el[1].get(w_attr, 1)), weight=0)
-        bpg.add_edge(xi, yi, capacity=float(sys.maxsize), weight=1)
-        bpg.add_edge(yi, 't', capacity=float(el[1].get(w_attr, 1)), weight=0)
+        bpg.add_edge('s', xi, capacity=int(el[1].get(w_attr, 1)), weight=0)
+        bpg.add_edge(xi, yi, capacity=int(sys.maxsize), weight=1)
+        bpg.add_edge(yi, 't', capacity=int(el[1].get(w_attr, 1)), weight=0)
 
         el_des = nx.descendants(dag, el[0])
         el_pred = nx.ancestors(dag, el[0])
@@ -77,8 +77,8 @@ def _get_pi_solution(split_graph):
         H = nx.DiGraph()
         H.add_nodes_from(split_graph)
         for ed in split_graph.edges(data=True):
-            Cxy = float(ed[2].get('capacity', sys.maxsize))
-            Axy = float(ed[2]['weight'])
+            Cxy = int(ed[2].get('capacity', sys.maxsize))
+            Axy = int(ed[2]['weight'])
             logger.debug(f"Found capacity and weight: {Axy}, {Cxy}")
             if (Axy == 0 and Cxy > 0):
                 H.add_edge(ed[0], ed[1], capacity=Cxy, weight=Axy)
